+++ src/general.ts
@@ -1,0 +1,7 @@
+/**
+ * The `identity` function which returns its argument unchanged.
+ * (Sometimes called `empty` or `zero`)
+ * @param x The value to return.
+ * @returns The value unchanged.
+ */
+export const id = <T>(x: T): T => x;