+++ tests/result.tests.ts
@@ -1,0 +1,6 @@
+import * as R from "../src/result"
+import { describe } from "vitest";
+
+describe.skip("result", () => {
+
+})